[finance_report_audit] refactor: retire apps/backend/tests/services/ — finish the test-file half of #1416
The #1416 package-model migration physically moved every source file out of
the now-deleted apps/backend/src/services/ into its owning package, but the
tests for that code never followed. 11 tracked files sat in
apps/backend/tests/services/, all importing already-package-owned modules
(src.extraction.*, src.reconciliation.*, src.reporting.*, src.runtime.*) —
none imported anything from the deleted src.services. 8 of them were the
literal `test=` proof path cited by 54 AC entries across
common/extraction/contract.py, common/audit/contract.py,
common/reconciliation/contract.py, docs/project/traceability-exceptions.md,
and common/meta/data/MANIFEST.yaml.

- 10 files git mv'd to their owning package's test dir, following whichever
  contract.py roadmap cites the test as proof (the authoritative ownership
  signal); import origin broke ties for the 3 files no roadmap cites.
- test_roi_value_type_adoption.py split 3-way (zero shared fixtures across
  its 3 unrelated test functions, verified by reading the full 82-line
  source): the reconciliation-tolerance test to
  tests/reconciliation/test_combination_tolerance_legacy_parity.py, the
  portfolio-allocation tests to
  tests/portfolio/test_allocation_value_share.py, and the reporting
  line-total test to tests/reporting/test_line_total_value_type.py.
- All 54 governance path-string references updated to match; the cassette
  test's self-referencing `make llm-record` docstring example fixed too.
- apps/backend/tests/services/ deleted (empty after the moves).
- Also removed apps/backend/tests/models/ and tests/auth/__pycache__/ — both
  100% untracked local build-artifact debris (the real auth tests already
  live at tests/identity/test_auth*.py).

Verified: the 5 destination package test dirs (1419 tests) pass from their
new homes; AC-traceability gate tests (108) confirm every edited `test=`
path resolves; full preflight (--tier=full --base main) green, 2130 tests
passed.
