+++ apps/backend/tests/audit/test_currency_resolution.py
