+++ apps/backend/tests/extraction/test_classification_migration.py
